+++ public/assets/JS/main.js
@@ -1,3 +1,5 @@
 import { populateFragments } from './fragments.js';
+import {loadSidebar} from "../JS/sidebar.js"
 
 populateFragments();
+loadSidebar();